backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.4
celery==5.4.0
websockets==12.0
PyJWT==2.8.0